@@ -1,65 +1,60 @@
-import java.text.SimpleDateFormat
-
 buildscript {
     ext {
-        springBootVersion = '1.5.9.RELEASE'
+        springBootVersion = '2.1.1.RELEASE'
     }
     repositories {
+        jcenter()
         mavenCentral()
     }
     dependencies {
         classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
     }
+
 }
 
-group 'uk.gov.justice.digital.hmpps'
-
 apply plugin: 'java'
-apply plugin: 'groovy'
-apply plugin: 'idea'
 apply plugin: 'org.springframework.boot'
-
-
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+apply plugin: 'io.spring.dependency-management'
+apply plugin: 'groovy'
 
 repositories {
     mavenCentral()
 }
 
-def applicationMainClass = 'uk.gov.justice.digital.hmpps.keyworker.KeyworkerApplication'
-def dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
-def dateOnlyFormat = new SimpleDateFormat("yyyy-MM-dd")
-def today = dateOnlyFormat.format(new Date())
-def applicationName = 'keyworker-api'
-version today
-
-jar {
-    baseName = applicationName
-    manifest {
-        attributes 'Implementation-Title': applicationName,
-                'Implementation-Version': System.getenv('CI') ? "${version}.${System.getenv('CIRCLE_BUILD_NUM')}" : version
+ext {
+    ext {
+        drivers = [
+                'chromeHeadless'
+        ]
     }
 }
 
+def today = java.time.Instant.now()
+group = 'uk.gov.justice.digital.hmpps'
+project.sourceCompatibility = 11
+project.targetCompatibility = 11
+
+def todayDate = java.time.LocalDate.now()
+def formatter = java.time.format.DateTimeFormatter.ofPattern("yyyy-MM-dd")
+version todayDate.format(formatter)
+
+
 springBoot {
-    mainClass = applicationMainClass
-
     buildInfo {
-        // Generate extra build info.
-        additionalProperties = [
-                by                   : System.properties['user.name'],
-                operatingSystem      : "${System.properties['os.name']} (${System.properties['os.version']})",
-                continuousIntegration: System.getenv('CI') ? true : false,
-                machine              : InetAddress.localHost.hostName,
-                time                 : dateFormat.format(new Date()),
-                name                 : applicationName,
-                version              : System.getenv('CI') ? "${version}.${System.getenv('CIRCLE_BUILD_NUM')}" : version
-        ]
+        properties {
+            artifact = rootProject.name
+            version = System.getenv('CI') ? "${version}.${System.getenv('CIRCLE_BUILD_NUM')}" : version
+            group = 'uk.gov.justice.digital.hmpps'
+            name = rootProject.name
+            time = today
+            additional = [
+                    by                   : System.properties['user.name'],
+                    operatingSystem      : "${System.properties['os.name']} (${System.properties['os.version']})",
+                    continuousIntegration: System.getenv('CI') ? true : false,
+                    machine              : InetAddress.localHost.hostName
+            ]
+        }
     }
-    executable = true
-
 }
 
 configurations {
