@@ -9,7 +9,7 @@ import java.util.List;
 
 @Component
 @Data
-@ConfigurationProperties(prefix="roleMigration")
+@ConfigurationProperties(prefix="role-migration")
 public class RoleMigrationConfiguration {
     private List<String> rolesToMatch = new ArrayList<>();
     private List<String> rolesToAssign = new ArrayList<>();
